Refactor current Review step
